+++ esphome/spectra_ls_system/rp2040_uart.h
@@ -1,6 +1,6 @@
 // Description: Legacy RP2040 UART hub shim header for Spectra LS system includes.
-// Version: 2026.06.21.1
-// Last updated: 2026-06-21
+// Version: 2026.07.31.1
+// Last updated: 2026-07-31
 
 #pragma once
 
@@ -116,7 +116,7 @@ class Rp2040UartHub : public Component, public uart::UARTDevice {
         break;
       }
       default:
-        ESP_LOGW("rp2040_uart", "unknown type=%u id=%u raw=%u ts=%u", type, id, raw_value, ts);
+        ESP_LOGW("rp2040_uart", "unknown type=%u id=%u raw=%u ts=%lu", type, id, raw_value, (unsigned long) ts);
         break;
     }
   }
